pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lc-admin-backend"
version = "0.1.0"
description = "Admin API for Legacy Cards"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic-settings>=2.5.0,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "PyJWT>=2.9.0,<3.0.0",
  "pyotp>=2.9.0,<3.0.0",
  "google-cloud-firestore>=2.21.0,<3.0.0",
  "google-cloud-bigquery>=3.31.0,<4.0.0",
  "google-cloud-storage>=2.18.2,<3.0.0",
  "cryptography>=46.0.7,<47.0.0",
  "openai>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "httpx>=0.27.0,<1.0.0",
  "ruff>=0.6.0,<1.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
